requirements.txt
# Python pakages required for main.py

pytube3